test: update tests according to transfer nft integration
