+++ apps/web/src/app/wallet/domain/models/nft.ts
@@ -1,0 +1,12 @@
+export interface Nft {
+  id?: string
+  token_id: string
+  name: string
+  description: string
+  url: string
+  code?: string
+  issuer?: string
+  resource?: string
+  transaction_hash: string
+  contract_address?: string
+}
