+++ apps/web/src/app/wallet/pages/left-assets/tabs/transfer-nfts/index.tsx
@@ -6,11 +6,11 @@ import { useForm } from 'react-hook-form'
 import { useToast } from 'src/app/core/hooks/use-toast'
 import { Toast } from 'src/app/core/services/toast'
 import { walletAddressFormSchema, WalletAddressFormValues } from 'src/app/wallet/components/wallet-address-form/schema'
+import { Nft } from 'src/app/wallet/domain/models/nft'
 import { getNfts, useGetNfts } from 'src/app/wallet/queries/use-get-nfts'
 import { getWallet } from 'src/app/wallet/queries/use-get-wallet'
 import { useTransfer } from 'src/app/wallet/queries/use-transfer'
 import { WalletPagesPath } from 'src/app/wallet/routes/types'
-import { Nft } from 'src/app/wallet/services/wallet/types'
 import { modalService } from 'src/components/organisms/modal/provider'
 import { c } from 'src/interfaces/cms/useContent'
 import { queryClient } from 'src/interfaces/query-client'
