+++ apps/web/src/app/wallet/pages/nfts/index.tsx
@@ -5,9 +5,9 @@ import { featureFlagsState } from 'src/app/core/helpers'
 
 import NftsTemplate from './template'
 import { ViewNftDrawer } from '../../components'
+import { Nft } from '../../domain/models/nft'
 import { useGetNfts } from '../../queries/use-get-nfts'
 import { WalletPagesPath } from '../../routes/types'
-import { Nft } from '../../services/wallet/types'
 
 export const Nfts = () => {
   const navigate = useNavigate()
